refactored for flask-sqlalchemy >=3.0.0
