+++ mslib/mscolab/app/__init__.py
@@ -24,12 +24,37 @@
     limitations under the License.
 """
 
+import os
+import mslib
 
-from flask_migrate import Migrate
-from mslib.mscolab.models import db
-from mslib.mscolab.server import _app as app
+from flask import Flask
+from mslib.mscolab.conf import mscolab_settings
+from mslib.mswms.gallery_builder import STATIC_LOCATION
+from mslib.utils import prefix_route
+
+
+DOCS_SERVER_PATH = os.path.dirname(os.path.abspath(mslib.__file__))
+# This can be used to set a location by SCRIPT_NAME for testing. e.g. export SCRIPT_NAME=/demo/
+SCRIPT_NAME = os.environ.get('SCRIPT_NAME', '/')
 
 # in memory database for testing
 # app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///'
+APP = Flask(__name__, template_folder=os.path.join(DOCS_SERVER_PATH, 'static', 'templates'), static_url_path="/static",
+            static_folder=STATIC_LOCATION)
+APP.config.from_object(__name__)
+APP.route = prefix_route(APP.route, SCRIPT_NAME)
 
-migrate = Migrate(app, db, render_as_batch=True)
+APP.config['MSCOLAB_DATA_DIR'] = mscolab_settings.MSCOLAB_DATA_DIR
+APP.config['SQLALCHEMY_DATABASE_URI'] = mscolab_settings.SQLALCHEMY_DB_URI
+APP.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+APP.config['UPLOAD_FOLDER'] = mscolab_settings.UPLOAD_FOLDER
+APP.config['MAX_CONTENT_LENGTH'] = mscolab_settings.MAX_UPLOAD_SIZE
+APP.config['SECRET_KEY'] = mscolab_settings.SECRET_KEY
+APP.config['SECURITY_PASSWORD_SALT'] = getattr(mscolab_settings, "SECURITY_PASSWORD_SALT", None)
+APP.config['MAIL_DEFAULT_SENDER'] = getattr(mscolab_settings, "MAIL_DEFAULT_SENDER", None)
+APP.config['MAIL_SERVER'] = getattr(mscolab_settings, "MAIL_SERVER", None)
+APP.config['MAIL_PORT'] = getattr(mscolab_settings, "MAIL_PORT", None)
+APP.config['MAIL_USERNAME'] = getattr(mscolab_settings, "MAIL_USERNAME", None)
+APP.config['MAIL_PASSWORD'] = getattr(mscolab_settings, "MAIL_PASSWORD", None)
+APP.config['MAIL_USE_TLS'] = getattr(mscolab_settings, "MAIL_USE_TLS", None)
+APP.config['MAIL_USE_SSL'] = getattr(mscolab_settings, "MAIL_USE_SSL", None)
