+++ mslib/utils/__init__.py
@@ -125,3 +125,16 @@ def conditional_decorator(dec, condition):
             return func
         return dec(func)
     return decorator
+
+
+def prefix_route(route_function, prefix='', mask='{0}{1}'):
+    '''
+    https://stackoverflow.com/questions/18967441/add-a-prefix-to-all-flask-routes/18969161#18969161
+    Defines a new route function with a prefix.
+    The mask argument is a `format string` formatted with, in that order:
+      prefix, route
+    '''
+    def newroute(route, *args, **kwargs):
+        ''' prefix route '''
+        return route_function(mask.format(prefix, route), *args, **kwargs)
+    return newroute
