+++ tests/_test_mscolab/test_file_manager.py
@@ -29,7 +29,7 @@ import os
 import pytest
 
 from mslib.mscolab.conf import mscolab_settings
-from mslib.mscolab.models import Operation, db
+from mslib.mscolab.models import Operation
 from mslib.mscolab.server import APP
 from mslib.mscolab.file_manager import FileManager
 from mslib.mscolab.seed import add_user, get_user
@@ -54,7 +54,6 @@ class Test_FileManager(TestCase):
 
     def setUp(self):
         handle_db_reset()
-        db.init_app(self.app)
         self.userdata = 'UV10@uv10', 'UV10', 'uv10'
         self.anotheruserdata = 'UV20@uv20', 'UV20', 'uv20'
         self.fm = FileManager(self.app.config["MSCOLAB_DATA_DIR"])
