+++ tests/_test_mscolab/test_seed.py
@@ -27,7 +27,7 @@
 from flask_testing import TestCase
 
 from mslib.mscolab.conf import mscolab_settings
-from mslib.mscolab.models import db, User, Operation
+from mslib.mscolab.models import User, Operation
 from mslib.mscolab.mscolab import handle_db_reset
 from mslib.mscolab.server import APP
 from mslib.mscolab.file_manager import FileManager
@@ -51,7 +51,6 @@ class Test_Seed(TestCase):
 
     def setUp(self):
         handle_db_reset()
-        db.init_app(self.app)
         self.fm = FileManager(self.app.config["MSCOLAB_DATA_DIR"])
         self.operation_name = "XYZ"
         self.description = "Template"
