+++ tests/_test_utils/test_index.py
@@ -26,7 +26,7 @@
 """
 
 from mslib import index
-
+from mslib.mscolab.server import APP as app
 
 def test_xstatic():
     assert index._xstatic('jquery') is not None
@@ -36,8 +36,6 @@ def test_xstatic():
 
 def test_app_loader():
     assert index.DOCS_SERVER_PATH.endswith('mslib')
-    app = index.app_loader(__name__)
-    assert app is not None
     with app.test_client() as c:
         response = c.get('/xstatic/bootstrap/css/bootstrap.css')
         assert response.status_code == 200
